Move store to its own directory
We'll want to put more vuex related stuff with it like mutation names.
