+++ src/main.ts
@@ -4,7 +4,7 @@ import Vue from 'vue';
 import './plugins/bootstrap-vue';
 import App from './App.vue';
 import router from './router';
-import store from './store';
+import store from './store/store';
 import './registerServiceWorker';
 
 Vue.config.productionTip = false;
